Move deprecated build methods to Factories
