+++ optaplanner-benchmark/src/main/java/org/optaplanner/benchmark/config/report/BenchmarkReportConfig.java
@@ -16,7 +16,6 @@
 
 package org.optaplanner.benchmark.config.report;
 
-import java.time.ZoneId;
 import java.util.Comparator;
 import java.util.Locale;
 
@@ -25,11 +24,6 @@ import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 
 import org.optaplanner.benchmark.config.ranking.SolverRankingType;
 import org.optaplanner.benchmark.impl.ranking.SolverRankingWeightFactory;
-import org.optaplanner.benchmark.impl.ranking.TotalRankSolverRankingWeightFactory;
-import org.optaplanner.benchmark.impl.ranking.TotalScoreSolverRankingComparator;
-import org.optaplanner.benchmark.impl.ranking.WorstScoreSolverRankingComparator;
-import org.optaplanner.benchmark.impl.report.BenchmarkReport;
-import org.optaplanner.benchmark.impl.result.PlannerBenchmarkResult;
 import org.optaplanner.benchmark.impl.result.SolverBenchmarkResult;
 import org.optaplanner.core.config.AbstractConfig;
 import org.optaplanner.core.config.util.ConfigUtils;
@@ -90,72 +84,8 @@ public class BenchmarkReportConfig extends AbstractConfig<BenchmarkReportConfig>
         this.solverRankingWeightFactoryClass = solverRankingWeightFactoryClass;
     }
 
-    // ************************************************************************
-    // Builder methods
-    // ************************************************************************
-
-    public BenchmarkReport buildBenchmarkReport(PlannerBenchmarkResult plannerBenchmark) {
-        BenchmarkReport benchmarkReport = new BenchmarkReport(plannerBenchmark);
-        benchmarkReport.setLocale(determineLocale());
-        benchmarkReport.setTimezoneId(ZoneId.systemDefault());
-        supplySolverRanking(benchmarkReport);
-        return benchmarkReport;
-    }
-
     public Locale determineLocale() {
-        return locale == null ? Locale.getDefault() : locale;
-    }
-
-    protected void supplySolverRanking(BenchmarkReport benchmarkReport) {
-        if (solverRankingType != null && solverRankingComparatorClass != null) {
-            throw new IllegalStateException("The PlannerBenchmark cannot have"
-                    + " a solverRankingType (" + solverRankingType
-                    + ") and a solverRankingComparatorClass (" + solverRankingComparatorClass.getName()
-                    + ") at the same time.");
-        } else if (solverRankingType != null && solverRankingWeightFactoryClass != null) {
-            throw new IllegalStateException("The PlannerBenchmark cannot have"
-                    + " a solverRankingType (" + solverRankingType
-                    + ") and a solverRankingWeightFactoryClass (" + solverRankingWeightFactoryClass.getName()
-                    + ") at the same time.");
-        } else if (solverRankingComparatorClass != null && solverRankingWeightFactoryClass != null) {
-            throw new IllegalStateException("The PlannerBenchmark cannot have"
-                    + " a solverRankingComparatorClass (" + solverRankingComparatorClass.getName()
-                    + ") and a solverRankingWeightFactoryClass (" + solverRankingWeightFactoryClass.getName()
-                    + ") at the same time.");
-        }
-        Comparator<SolverBenchmarkResult> solverRankingComparator = null;
-        SolverRankingWeightFactory solverRankingWeightFactory = null;
-        if (solverRankingType != null) {
-            switch (solverRankingType) {
-                case TOTAL_SCORE:
-                    solverRankingComparator = new TotalScoreSolverRankingComparator();
-                    break;
-                case WORST_SCORE:
-                    solverRankingComparator = new WorstScoreSolverRankingComparator();
-                    break;
-                case TOTAL_RANKING:
-                    solverRankingWeightFactory = new TotalRankSolverRankingWeightFactory();
-                    break;
-                default:
-                    throw new IllegalStateException("The solverRankingType ("
-                            + solverRankingType + ") is not implemented.");
-            }
-        }
-        if (solverRankingComparatorClass != null) {
-            solverRankingComparator = ConfigUtils.newInstance(this,
-                    "solverRankingComparatorClass", solverRankingComparatorClass);
-        }
-        if (solverRankingWeightFactoryClass != null) {
-            solverRankingWeightFactory = ConfigUtils.newInstance(this,
-                    "solverRankingWeightFactoryClass", solverRankingWeightFactoryClass);
-        }
-        if (solverRankingComparator != null) {
-            benchmarkReport.setSolverRankingComparator(solverRankingComparator);
-        } else if (solverRankingWeightFactory != null) {
-            benchmarkReport.setSolverRankingWeightFactory(solverRankingWeightFactory);
-        } else {
-            benchmarkReport.setSolverRankingComparator(new TotalScoreSolverRankingComparator());
-        }
+        return getLocale() == null ? Locale.getDefault() : getLocale();
     }
 
     @Override
