+++ optaplanner-benchmark/src/main/java/org/optaplanner/benchmark/impl/loader/ProblemProvider.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,13 +18,12 @@ package org.optaplanner.benchmark.impl.loader;
 
 import javax.xml.bind.annotation.XmlSeeAlso;
 
-import org.optaplanner.benchmark.config.ProblemBenchmarksConfig;
 import org.optaplanner.benchmark.impl.result.SubSingleBenchmarkResult;
 import org.optaplanner.core.api.domain.solution.PlanningSolution;
 
 /**
  * Subclasses need to implement {@link Object#equals(Object) equals()} and {@link Object#hashCode() hashCode()}
- * which are used by {@link ProblemBenchmarksConfig#buildProblemBenchmarkList}.
+ * which are used by {@link org.optaplanner.benchmark.impl.ProblemBenchmarksFactory#buildProblemBenchmarkList}.
  *
  * @param <Solution_> the solution type, the class with the {@link PlanningSolution} annotation
  */
