+++ optaplanner-benchmark/src/test/java/org/optaplanner/benchmark/config/PlannerBenchmarkConfigTest.java
@@ -18,8 +18,6 @@ package org.optaplanner.benchmark.config;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-import static org.assertj.core.api.Assertions.assertThatIllegalStateException;
 
 import java.io.IOException;
 import java.io.InputStreamReader;
@@ -28,10 +26,6 @@ import java.io.StringReader;
 import java.io.StringWriter;
 import java.io.Writer;
 import java.nio.charset.StandardCharsets;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.TreeSet;
 
 import org.apache.commons.io.IOUtils;
 import org.junit.jupiter.api.Test;
@@ -46,123 +40,6 @@ class PlannerBenchmarkConfigTest {
 
     private static final String TEST_PLANNER_BENCHMARK_CONFIG_WITH_NAMESPACE = "testBenchmarkConfigWithNamespace.xml";
     private static final String TEST_PLANNER_BENCHMARK_CONFIG_WITHOUT_NAMESPACE = "testBenchmarkConfigWithoutNamespace.xml";
-
-    @Test
-    void validNameWithUnderscoreAndSpace() {
-        PlannerBenchmarkConfig config = new PlannerBenchmarkConfig();
-        config.setName("Valid_name with space_and_underscore");
-        config.setSolverBenchmarkConfigList(Collections.singletonList(new SolverBenchmarkConfig()));
-        config.validate();
-    }
-
-    @Test
-    void validNameWithJapanese() {
-        PlannerBenchmarkConfig config = new PlannerBenchmarkConfig();
-        config.setName("Valid name (有効名 in Japanese)");
-        config.setSolverBenchmarkConfigList(Collections.singletonList(new SolverBenchmarkConfig()));
-        config.validate();
-    }
-
-    @Test
-    void invalidNameWithSlash() {
-        PlannerBenchmarkConfig config = new PlannerBenchmarkConfig();
-        config.setName("slash/name");
-        config.setSolverBenchmarkConfigList(Collections.singletonList(new SolverBenchmarkConfig()));
-        assertThatIllegalStateException().isThrownBy(config::validate);
-    }
-
-    @Test
-    void invalidNameWithSuffixWhitespace() {
-        PlannerBenchmarkConfig config = new PlannerBenchmarkConfig();
-        config.setName("Suffixed with space ");
-        config.setSolverBenchmarkConfigList(Collections.singletonList(new SolverBenchmarkConfig()));
-        assertThatIllegalStateException().isThrownBy(config::validate);
-    }
-
-    @Test
-    void invalidNameWithPrefixWhitespace() {
-        PlannerBenchmarkConfig config = new PlannerBenchmarkConfig();
-        config.setName(" prefixed with space");
-        config.setSolverBenchmarkConfigList(Collections.singletonList(new SolverBenchmarkConfig()));
-        assertThatIllegalStateException().isThrownBy(config::validate);
-    }
-
-    @Test
-    void noSolverConfigs() {
-        PlannerBenchmarkConfig config = new PlannerBenchmarkConfig();
-        config.setSolverBenchmarkConfigList(null);
-        config.setSolverBenchmarkBluePrintConfigList(null);
-        assertThatIllegalArgumentException().isThrownBy(config::validate);
-    }
-
-    @Test
-    void nonUniqueSolverConfigName() {
-        PlannerBenchmarkConfig config = new PlannerBenchmarkConfig();
-        final String sbcName = "x";
-        SolverBenchmarkConfig sbc1 = new SolverBenchmarkConfig();
-        sbc1.setName(sbcName);
-        SolverBenchmarkConfig sbc2 = new SolverBenchmarkConfig();
-        sbc2.setName(sbcName);
-        config.setSolverBenchmarkConfigList(Arrays.asList(sbc1, sbc2));
-        assertThatIllegalStateException().isThrownBy(config::generateSolverBenchmarkConfigNames);
-    }
-
-    @Test
-    void uniqueNamesGenerated() {
-        PlannerBenchmarkConfig config = new PlannerBenchmarkConfig();
-        SolverBenchmarkConfig sbc1 = new SolverBenchmarkConfig();
-        SolverBenchmarkConfig sbc2 = new SolverBenchmarkConfig();
-        SolverBenchmarkConfig sbc3 = new SolverBenchmarkConfig();
-        sbc3.setName("Config_1");
-        List<SolverBenchmarkConfig> configs = Arrays.asList(sbc1, sbc2, sbc3);
-        config.setSolverBenchmarkConfigList(configs);
-        config.generateSolverBenchmarkConfigNames();
-        assertThat(sbc3.getName()).isEqualTo("Config_1");
-        TreeSet<String> names = new TreeSet<>();
-        for (SolverBenchmarkConfig sc : configs) {
-            names.add(sc.getName());
-        }
-        for (int i = 0; i < configs.size(); i++) {
-            assertThat(names).contains("Config_" + i);
-        }
-    }
-
-    @Test
-    void resolveParallelBenchmarkCountAutomatically() {
-        PlannerBenchmarkConfig config = new PlannerBenchmarkConfig();
-        assertThat(config.resolveParallelBenchmarkCountAutomatically(-1)).isEqualTo(1);
-        assertThat(config.resolveParallelBenchmarkCountAutomatically(0)).isEqualTo(1);
-        assertThat(config.resolveParallelBenchmarkCountAutomatically(1)).isEqualTo(1);
-        assertThat(config.resolveParallelBenchmarkCountAutomatically(2)).isEqualTo(1);
-        assertThat(config.resolveParallelBenchmarkCountAutomatically(3)).isEqualTo(2);
-        assertThat(config.resolveParallelBenchmarkCountAutomatically(4)).isEqualTo(2);
-        assertThat(config.resolveParallelBenchmarkCountAutomatically(5)).isEqualTo(3);
-        assertThat(config.resolveParallelBenchmarkCountAutomatically(6)).isEqualTo(4);
-        assertThat(config.resolveParallelBenchmarkCountAutomatically(17)).isEqualTo(9);
-    }
-
-    @Test
-    void parallelBenchmarkDisabledByDefault() {
-        PlannerBenchmarkConfig config = new PlannerBenchmarkConfig();
-        assertThat(config.resolveParallelBenchmarkCount()).isEqualTo(1);
-    }
-
-    @Test
-    void resolvedParallelBenchmarkCountNegative() {
-        PlannerBenchmarkConfig config = new PlannerBenchmarkConfig();
-        config.setParallelBenchmarkCount("-1");
-        assertThatIllegalArgumentException().isThrownBy(config::resolveParallelBenchmarkCount);
-    }
-
-    @Test
-    void calculateWarmUpTimeMillisSpentLimit() {
-        PlannerBenchmarkConfig config = new PlannerBenchmarkConfig();
-        config.setWarmUpHoursSpentLimit(1L);
-        config.setWarmUpMinutesSpentLimit(2L);
-        config.setWarmUpSecondsSpentLimit(5L);
-        config.setWarmUpMillisecondsSpentLimit(753L);
-        assertThat(config.calculateWarmUpTimeMillisSpentLimit()).isEqualTo(3_725_753L);
-    }
 
     @ParameterizedTest
     @ValueSource(strings = { TEST_PLANNER_BENCHMARK_CONFIG_WITHOUT_NAMESPACE, TEST_PLANNER_BENCHMARK_CONFIG_WITH_NAMESPACE })
