+++ optaplanner-benchmark/src/test/java/org/optaplanner/benchmark/impl/SolverBenchmarkFactoryTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,20 +14,21 @@
  * limitations under the License.
  */
 
-package org.optaplanner.benchmark.config;
+package org.optaplanner.benchmark.impl;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalStateException;
 
 import org.junit.jupiter.api.Test;
+import org.optaplanner.benchmark.config.SolverBenchmarkConfig;
 
-public class SolverBenchmarkConfigTest {
+public class SolverBenchmarkFactoryTest {
 
     @Test
     public void validNameWithUnderscoreAndSpace() {
         SolverBenchmarkConfig config = new SolverBenchmarkConfig();
         config.setName("Valid_name with space_and_underscore");
         config.setSubSingleCount(1);
-        config.validate();
+        validateConfig(config);
     }
 
     @Test
@@ -35,7 +36,7 @@ public class SolverBenchmarkConfigTest {
         SolverBenchmarkConfig config = new SolverBenchmarkConfig();
         config.setName("Valid name (有効名 in Japanese)");
         config.setSubSingleCount(1);
-        config.validate();
+        validateConfig(config);
     }
 
     @Test
@@ -43,7 +44,7 @@ public class SolverBenchmarkConfigTest {
         SolverBenchmarkConfig config = new SolverBenchmarkConfig();
         config.setName("slash/name");
         config.setSubSingleCount(1);
-        assertThatIllegalStateException().isThrownBy(config::validate);
+        assertThatIllegalStateException().isThrownBy(() -> validateConfig(config));
     }
 
     @Test
@@ -51,7 +52,7 @@ public class SolverBenchmarkConfigTest {
         SolverBenchmarkConfig config = new SolverBenchmarkConfig();
         config.setName("Suffixed with space ");
         config.setSubSingleCount(1);
-        assertThatIllegalStateException().isThrownBy(config::validate);
+        assertThatIllegalStateException().isThrownBy(() -> validateConfig(config));
     }
 
     @Test
@@ -59,7 +60,7 @@ public class SolverBenchmarkConfigTest {
         SolverBenchmarkConfig config = new SolverBenchmarkConfig();
         config.setName(" prefixed with space");
         config.setSubSingleCount(1);
-        assertThatIllegalStateException().isThrownBy(config::validate);
+        assertThatIllegalStateException().isThrownBy(() -> validateConfig(config));
     }
 
     @Test
@@ -67,7 +68,7 @@ public class SolverBenchmarkConfigTest {
         SolverBenchmarkConfig config = new SolverBenchmarkConfig();
         config.setName("name");
         config.setSubSingleCount(2);
-        config.validate();
+        validateConfig(config);
     }
 
     @Test
@@ -75,7 +76,7 @@ public class SolverBenchmarkConfigTest {
         SolverBenchmarkConfig config = new SolverBenchmarkConfig();
         config.setName("name");
         config.setSubSingleCount(null);
-        config.validate();
+        validateConfig(config);
     }
 
     @Test
@@ -83,7 +84,11 @@ public class SolverBenchmarkConfigTest {
         SolverBenchmarkConfig config = new SolverBenchmarkConfig();
         config.setName("name");
         config.setSubSingleCount(0);
-        assertThatIllegalStateException().isThrownBy(config::validate);
+        assertThatIllegalStateException().isThrownBy(() -> validateConfig(config));
     }
 
+    private void validateConfig(SolverBenchmarkConfig config) {
+        SolverBenchmarkFactory solverBenchmarkFactory = new SolverBenchmarkFactory(config);
+        solverBenchmarkFactory.validate();
+    }
 }
